adding resource into client-side

+++ pages/api/resource/[id].js
@@ -1,9 +1,10 @@
 import nc from 'next-connect'
-import dbConnect from '../../../../utils/db'
-import { isAuth } from '../../../../utils/auth'
+import dbConnect from '../../../utils/db'
+import { isAuth } from '../../../utils/auth'
 import fileUpload from 'express-fileupload'
-import { upload, deleteFile } from '../../../../utils/fileManager'
-import Resource from '../../../../models/Resource'
+import { upload, deleteFile } from '../../../utils/fileManager'
+import Resource from '../../../models/Resource'
+import AssignSubject from '../../../models/AssignSubject'
 export const config = { api: { bodyParser: false } }
 
 const handler = nc()
@@ -13,30 +14,51 @@ handler.use(isAuth)
 
 handler.put(async (req, res) => {
   await dbConnect()
-  const { courseType, course, subject, instructor, semester, shift, isActive } =
-    req.body
-
+  const {
+    courseType,
+    course,
+    description,
+    subject,
+    semester,
+    shift,
+    isActive,
+  } = req.body
   const _id = req.query.id
   const file = req.files && req.files.file
+  const instructor = req.user.instructor
 
   const obj = await Resource.findById(_id)
 
+  const assign = await AssignSubject.find({
+    subject,
+    instructor,
+    shift,
+    isActive: true,
+  })
+
+  if (assign.length === 0)
+    return res.status(400).send('You do not belong the subject you selected')
+
   if (obj) {
     if (file) {
-      if (obj && obj.file) {
-        deleteFile({
-          pathName: obj.file.filePath,
-        })
-      }
-
       const document = await upload({
         fileName: file,
-        fileType: 'image',
-        pathName: 'resource',
+        fileType: 'file',
+        pathName: 'recourse',
       })
+
+      if (document) {
+        if (obj && obj.file) {
+          deleteFile({
+            pathName: obj.file.filePath,
+          })
+        }
+      }
+
       obj.isActive = isActive
       obj.courseType = courseType
       obj.course = course
+      obj.description = description
       obj.subject = subject
       obj.instructor = instructor
       obj.semester = semester
@@ -55,6 +77,7 @@ handler.put(async (req, res) => {
       obj.course = course
       obj.subject = subject
       obj.instructor = instructor
+      obj.description = description
       obj.semester = semester
       obj.shift = shift
       await obj.save()
